Drop minimum SDK for running the app.
This removes the basic restriction on running the sample application on
older devices. This allows older systems with geofence only support to
run the app.

In order to set the older SDK we need to tell the app to ignore what is
in the Google Play lib. The campaign kit library code checks the SDK
before enabling certain features, such as geofences and BLE support.