Use new Campaign Kit manager factory.
This updates the creation of the Campaign Kit manager to use the new
designated factory: `getInstance(Context, Configuration)`. With the
0.6.0 release the old constructor became deprecated. This is in
preparation for a future design change where it will be possible to have
more than one instance of a Campaign Kit manager.

Since the app will now be responsible for managing the singleton
instance of the manager we need to ensure we synchronize our creation.
While this is not necessary inside an `Application` subclass it will be
necessary inside an `Activity` or other Android/Java component. We know
people tend to just copy code so we are including it here to allow for
easy reference for others.