Request permissions for Android 6
This implements the basic flow for requesting permissions in Android 6.

The only permission we need to request is access to location services.
This is necessary so we can detect beacons in the background - via
`ACCESS_COARSE_LOCATION`. Additionally, Google Play Service needs access
to `ACCESS_FINE_LOCATION` for geofence triggering.

This follows the flow from the Google Android tutorials. For more
details see:

- https://developer.android.com/training/permissions/index.html
- https://www.google.com/design/spec/patterns/permissions.html
- https://www.youtube.com/watch?v=C8lUdPVSzDk
- https://www.youtube.com/watch?v=iZqDdvhTZj0

Since the permissions we need are critical. We ask upfront. As this is a
sample app it may, or may not, be clear why we need the permissions.
Since this is only a reference we are no providing any on boarding and
defer the explanation until necessary.

However, because the permissions we require are critical we move the
check from `onCreate` to `onStart`. This way, whenever the user
re-foregrounds the app we verify that we still have the necessary
permissions.