Update Google Play Service check for 8.3.0
This replaces the previous logic for checking for Google Play Service
with the new logic as suggested for version 8.3.0. Since we are also
targeting Android 6 this updates the flow so that we only attempt to
check for Google Play Services _if_ we have been given permission to use
location services. There's no point in checking if we are going to be
denied the one feature we need.