Add product flavors to support legacy testing
Gradle and Android Studio have supported product flavors for a while.
This updates the build to define flavors based on Android API and Google
Play services combinations. This allows easier testing of SDK builds
for drop-in-replacement for legacy apps.

The product flavors are setup to all use the same application ID by
default. This allows a developer to focus on a single API product at a
time without having to worry about uninstalling other versions of the
app. With the app being replaced when running different product flavors
this adds extra flavor specific information to the builds. The app name
and title now include a suffix using the build flavor. Similarly, the
log tags also have the flavor included in them. However, if a developer
wants to test multiple product flavors at the same time an application
ID suffix is included in the build, but commented out by default. To run
multiple apps at the same time the developer simply needs to uncomment
these lines.

In order to support multiple Google Play services this splits the Google
Play specific API calls into a singleton utility class. Unfortunately,
unlike manifests and resource files source files must be unique between
the product flavor and the main source set. To get around this and
continue to share code this sets up symlinks accordingly. The
authoritative source file is in the lowest API level which supports it.
This way it's clear that changes weren't made in new versions and if a
change is made the source can be copied accordingly.