Fix product flavor specific compile SDK version
Even though gradle / Android Studio doesn't complain when
`compileSdkVersion` is included within a product flavor it doesn't do
what you would expect. Instead of setting the compile version for just
that product flavor it sets it for all product flavors; just as if it
had been declared outside of the config block. In effect the last
declared value "wins", so in this case it would be the value specified
by the `api26` product flavor.

It makes sense that Android Studio, and gradle, wouldn't expect someone
to try to do this. For a real app you would never do this. You _want_ to
compile against a single SDK then simply have the appropriate version
checks to toggle the various calls. This even allows you to target
different versions in the product flavors without having to change how
things compile. This greatly reduces the amount of extra
handling required for cross flavor support.

Gradle doesn't make it easy for you to figure out what tasks were
requested to run. Thankfully, Android Studio pre-injects some properties
into the build when running the app. We leverage this to determine the
appropriate compile SDK to use. However, this does mean adding or change
reference app code may work on one build variant but fail to compile on
another. And this does require us to jump through a few more hoops at
times do handle the differences in SDKs.

Now that things are compiling against the proper base SDKs we need to
fix how permissions are handled. I had wondered why this wasn't
necessary in the previous commit and now I know! This takes the same
general approach as the Google Play utility class. It's a stateless
singleton to avoid worry about threading issues. In order to support
allowing the app to handle non-permission logic we adjust the API to
return the feature set instead of handling it with the permission check.

Finally, we have to add back the source compatibility we had previously
in order to compile against Android 19. Android 19 was complaining about
the diamond operators on collections.